# Conflicts:
#	API/src/main/java/org/sikuli/script/Finder2.java

+++ API/src/main/java/org/sikuli/script/Pattern.java
@@ -18,7 +18,7 @@ public class Pattern {
   static RunTime runTime = RunTime.get();
 
   private Image image = null;
-  private float similarity = (float) Settings.MinSimilarity;
+  private double similarity = Settings.MinSimilarity;
   private Location offset = new Location(0, 0);
   private int waitAfter = 0;
   private boolean imagePattern = false;
@@ -176,7 +176,7 @@ public class Pattern {
    * @param sim value 0 to 1
    * @return the Pattern object itself
    */
-  public Pattern similar(float sim) {
+  public Pattern similar(double sim) {
     similarity = sim;
     return this;
   }
@@ -187,7 +187,7 @@ public class Pattern {
    * @return the Pattern object itself
    */
   public Pattern exact() {
-    similarity = 0.99f;
+    similarity = 0.99;
     return this;
   }
 
@@ -195,7 +195,7 @@ public class Pattern {
    *
    * @return the current minimum similarity
    */
-  public float getSimilar() {
+  public double getSimilar() {
     return this.similarity;
   }
